requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
asgiref
sqlparse
psycopg2-binary
python-dotenv